requirements.txt
requests
beautifulsoup4
pillow
pytesseract
psycopg2-binary
regex
python-dateutil
